backend/requirements.txt
Flask==2.0.1 
Flask-CORS==3.0.10 
Pillow==9.5.0 
numpy==1.23.5
Werkzeug==2.0.3